src/presentation/controllers/loud-survey-result-controller.ts

import { forbidden, serverError, Ok } from '@/presentation/helpers/HttpHelpers'
import { InvalidParamError } from '@/presentation/errors'

import { Controller, HttpResponse } from '@/presentation/protocols'
import { CheckSurveyById, LoadSurveyResult } from '@/domain'

export class LoadSurveyResultController implements Controller {
  constructor (
    private readonly checkSurveyById: CheckSurveyById,
    private readonly loadSurveyResult: LoadSurveyResult
  ) {}

  async handle (request: LoadSurveyResultController.Request): Promise<HttpResponse> {
    try {
      const { surveyId } = request
      const exists = await this.checkSurveyById.checkById(surveyId)
      if (!exists) {
        return forbidden(new InvalidParamError('surveyId'))
      }
      const surveyResult = await this.loadSurveyResult.load(surveyId, request.accountId)
      return Ok(surveyResult)
    } catch (error) {
      return serverError(error)
    }
  }
}

export namespace LoadSurveyResultController {
  export type Request = {
    surveyId: string
    accountId: string
  }
}
